requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
django==5.1.2
django-ninja==1.3.0
pip==24.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
python-decouple==3.8
sqlparse==0.5.1
typing-extensions==4.12.2
